Add query string parameter for forcing alpha clip on uploaded media
This small change would force alpha clip (instead of blend) for uploaded images, to try resolving a bug in rendering stacked transparent images, as documented in https://github.com/mozilla/hubs/issues/5297

There's some uncertainty as to whether this affects .glb files which have alpha blend (for example i am adding semitransparent windows to some models) and whether that will cause problems for such scenes. 

Overall I'd like a little more control or transparency (no pun) on which uploaded media get which alpha behavior, so this is a very minimally invasive way to try resolving this bug. There's suggestion of later having a scene setting but I think it's too early for that. 

Thank you and many thanks to @Exairnous for a ton of help understanding and tracing this!