Resize huge textures for iOS devices
Browsers on iOS devices can crash if huge or many textures are
used in a room. Perhaps huge VRAM usage may cause it. We already
reported it to the webkit devs.

Until the root issue is fixed on iOS end, we resize huge texture
images in the low material quality mode as workaround. It isn't
a perfect solution but it should mitigate the problem.

Also disable image bitmap for iOS because texture image resize +
image bitmap can still even cause the problem. Disabling it
seemt to further mitigate the problem.