Fix darkened video-texture-target
This commit fixes darkened video-texture-target problem
caused when upgrading our Three.js to r141.

Refer to "Fix darkened camera view" commit for the details
because the root issue is same.

The previous same purpose commit was reverted because
it didn't work correctly because of losing
material.onBeforeRender when cloning materials.
This commit resolves it by copying the function.