Add LightTag and call Three.js light.dispose when removed
Add LightTag component that must be set for all lights. Call
Three.js light.dispose() when removed in remove-object3D-system.

This new tag and flow provides a proper light resource disposing.

Currently light.dispose() is called right after DirectionalLight
creation in inflator/directional-light if light.shadow.map is
non-null but it seems a bug because light.shadow.map should be
null when light is created. This commit also fixes it by removing
the lines. Light will be properly disposed when removed with
the change.