Implement rotate and scale button functions with bitecs
This commit adds the implementations of Object menu
rotate and scale button functions.

Note that this commit focuses on making them just functional
with minimal change.

The implementation relys on the old systems to save
development time. It should be rewritten more elegantly to
fit more to bitECS.

The main logic of src/components/scale-button is exposed
to be reused from the new and old systems.