bitECS object list support
This commit adds bitECS object list support.

**Basic Strategy**

Reuse the existing A-Frame based code as much as possible for now.

In the related functions that take A-Frame element, take Object3D
instead. Object3D has a reference to A-Frame element in the A-Frame
based implementation and has a reference to Entity ID in the bitECS.
The functions process with A-Frame element or Entity ID depending on
whether new loader enabled.

Explicitly use shouldUseNewLoader() and/or make two functions, one
for A-Frame based and another one for bitECS based implementation,
if different logics are needed between A-Frame based and bitECS
based implementations. Duplicated codes may not be perfectly
removed but it would be simpler to follow the code rather than
the new loader pretends the old one. And it would be easier to edit
the code when we will get rid of A-Frame.

**Changes**

- Introduce MediaInfo component and save url and content type
  into it when loading media in media-loader. The info is used
  in the object list
- Fire listed_media_changed event when media is loaded via
  media-loader and when the loaded media entity is removed
- Take Object3D instead of A-Frame element in the related functions
- Use shouldUseNewLoader() and/or make two separated functions
  for A-Frame and bitECS where the different logics are needed

**Future TODOs**

- Support avatars
- Support Pinning