bitECS: Take ownership when rotatin or scaling object
**Problem**

With new loader enabled, objects are not rotated or scaled
even if you attempt to rotate or scale objects that you don't
have ownership. You first need to click objects to take
ownership and then you can rotate or scale.

https://github.com/mozilla/hubs/issues/6199

But it is not a expected limitation. You should be able to
rotate or scale objects without taking ownership with
other operations beforehand.

**Root issue**

Rotate and Scale with Object menu don't take ownership.

**Solution**

Take ownership when starting rotate or scale.