Rename New Loader
This commit renames "New Loader" to "bitECS based Client"

Background:

The name "New Loader" may not really make sense to users
because the new code is not only related to loading stuffs
but it affects the Hubs Client core design and implementation.

It would be better to rename "New Loader" where users can see
the name, although we may keep internally using "New Loader"
as a project name.

Changes:
* Rename "New Loader" to "bitECS based Client" where it's exposed
  to users
* Add description for the toggle UI
* Remove outdated comments