Fix Browse button layout
**Problem**

Browse button in Custom Object upload dialog displays weirdly
in portrait mode on iOS. It unexpectedly covers almost the entire
URL area in portrait mode.

**Solution**

The issue seems to disappear by removing the "width: 100%" from
the parent element style. The root issue might be in iOS,
but we choose this workaround for now.