Remove Mozilla references and links.
Update Mozilla links everywhere to the Hubs-Foundation versions, or to placeholders until suitable documents can be recreated.  Placeholders are marked with (TBD) where possible.
Update various documentation files to include missing repositories, remove outdated information, and be more in line with the current state of things.  More work on this is needed.
References to Mozilla's name are either removed or changed to refer to Hubs, its community, or the Hubs Foundation.

Note: links in .env files and habitat files are unchanged to ensure nothing breaks.  The same for references to mozillareality.