Revert 699187a537b30d2b67dbe059d44311ceaebcefa0 and part of 1567f224c1db251e9540626211e68b4dea06b32e
Fixes https://github.com/Hubs-Foundation/hubs/issues/6496

Chromium version 128 introduced a bug in Hubs that was fixed by
https://github.com/Hubs-Foundation/hubs/pull/6495 but PR#6495 seems
to have introduced the bug detailed in issue#6496 for all browsers.
Now Chromium has fixed an issue on their end as detailed here:
https://issues.chromium.org/issues/361823993 which should fix
the original bug that PR#6495 addressed in Hubs, which allows us
to revert the fix on the Hubs side (PR#6495) and so fix the
issue#6496 bug.  However, this doesn't do anything for the
potentially deeper issues in the system that were possibly exposed
with PR#6495.