Force the repository name to be lowercase in the custom docker workflow
Converts the repository name to lowercase before storing it as an environment variable for use when connecting to the docker repository.

Some (or possibly all?) container registries don't allow their repository names to be capitalized, but the workflow gets the docker repository name from the GitHub repository name (which is allowed to be capitalized), so if the GitHub repository name is capitalized, this can prevent the workflow from working.